Fixed some UB

+++ common/logic/include/base/Dice.hpp
@@ -14,8 +14,8 @@ class DiceInterface {
     static const std::set<uint8_t> valid_dice_;
     
  public:
-    virtual uint8_t roll(uint8_t die) const = 0;
-    virtual std::vector<uint8_t> roll(uint8_t die, size_t num) const = 0;
+    virtual uint16_t roll(uint8_t die) const = 0;
+    virtual std::vector<uint16_t> roll(uint8_t die, size_t num) const = 0;
 
     virtual bool isValid(uint8_t die) const {
         return valid_dice_.contains(die);
@@ -41,8 +41,8 @@ class Dice : public DiceInterface {
 
     static constexpr uint8_t D20 = 20;
 
-    uint8_t roll(uint8_t die) const override;
+    uint16_t roll(uint8_t die) const override;
     std::unique_ptr<DiceInterface> clone() const override;
-    std::vector<uint8_t> roll(uint8_t die, size_t num) const override;
+    std::vector<uint16_t> roll(uint8_t die, size_t num) const override;
 };
 }  // namespace DnD
